Split widget: health bar
