+++ extension/devtools/app/lib/main.dart
@@ -29,6 +29,7 @@ part 'widgets/glass_pill.dart';
 part 'widgets/insight_card.dart';
 part 'widgets/insight_row.dart';
 part 'widgets/health_card.dart';
+part 'widgets/health_bar.dart';
 part 'widgets/metric_tile.dart';
 part 'widgets/panel_placeholder.dart';
 part 'widgets/panel_scroll_view.dart';
@@ -3206,69 +3207,6 @@ class _PanelInfo {
   final String title;
   final String description;
   final List<String> bullets;
-}
-
-class _HealthBar extends StatelessWidget {
-  const _HealthBar({
-    required this.label,
-    required this.value,
-    required this.progress,
-    required this.color,
-  });
-
-  final String label;
-  final String value;
-  final double progress;
-  final Color color;
-
-  @override
-  Widget build(BuildContext context) {
-    final textTheme = Theme.of(context).textTheme;
-    final clamped = progress.clamp(0.0, 1.0);
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Row(
-          children: [
-            Expanded(child: Text(label, style: textTheme.bodySmall)),
-            Text(value, style: textTheme.bodySmall),
-          ],
-        ),
-        const SizedBox(height: 6),
-        LayoutBuilder(
-          builder: (context, constraints) {
-            return Stack(
-              children: [
-                Container(
-                  height: 8,
-                  decoration: BoxDecoration(
-                    color: Theme.of(
-                      context,
-                    ).colorScheme.onSurface.withValues(alpha: 0.08),
-                    borderRadius: BorderRadius.circular(999),
-                  ),
-                ),
-                Container(
-                  height: 8,
-                  width: constraints.maxWidth * clamped,
-                  decoration: BoxDecoration(
-                    color: color,
-                    borderRadius: BorderRadius.circular(999),
-                    boxShadow: [
-                      BoxShadow(
-                        color: color.withValues(alpha: 0.35),
-                        blurRadius: 8,
-                      ),
-                    ],
-                  ),
-                ),
-              ],
-            );
-          },
-        ),
-      ],
-    );
-  }
 }
 
 class _ChartPlaceholder extends StatelessWidget {
